Python 3 long() to int()

+++ app.py
@@ -44,7 +44,7 @@ def status():
     filter_date = datetime.utcnow()
     guid = None
     if request.args is not None and 'max_created_at' in request.args:
-        filter_date = datetime.fromtimestamp(long(request.args['max_created_at']) / 1000.0)
+        filter_date = datetime.fromtimestamp(int(request.args['max_created_at']) / 1000.0)
     results = Status.query.filter(Status.created_at <= filter_date)
     if request.args is not None and 'userguid' in request.args:
         guid = request.args['userguid']
